base: remove Base repr as object name
Having the Base object repr be the name itself was causing confusion, as
the name and the object were in some cases being used interchangeably,
with no type checking.  By allowing the repr to be more descriptive (as well
as not using the name hash as the object hash from the previous commit) we
eliminate that ambiguity.